parlant-colosseo/requirements.txt
parlant==3.0.3
fastapi==0.115.12
uvicorn
pydantic
httpx
python-multipart
jinja2==3.1.6
aiofiles==24.1.0
PyYAML==6.0.1
